requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
distlib==0.3.7
filelock==3.12.2
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
Pillow==10.0.0
platformdirs==3.10.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.0
soupsieve==2.5
speedtest-cli==2.1.3
ttkbootstrap==1.10.1
urllib3==2.0.4
virtualenv==20.24.3
wikipedia==1.4.0
